+++ packages/datadog_session_replay/lib/src/capture/recorder.dart
@@ -128,11 +128,15 @@ class SessionReplayRecorder {
     DatadogTimeProvider timeProvider = const DefaultTimeProvider(),
     required TreeCapturePrivacy defaultCapturePrivacy,
     required TouchPrivacyLevel touchPrivacyLevel,
+    ImageDownscaling imageDownscaling = ImageDownscaling.disabled,
+    InternalLogger? internalLogger,
   }) : this._(
           KeyGenerator(),
           timeProvider,
           defaultCapturePrivacy,
           touchPrivacyLevel,
+          imageDownscaling,
+          internalLogger,
         );
 
   SessionReplayRecorder._(
@@ -140,13 +144,19 @@ class SessionReplayRecorder {
     this._timeProvider,
     this._defaultTreeCapturePrivacy,
     this._touchPrivacyLevel,
+    ImageDownscaling imageDownscaling,
+    InternalLogger? internalLogger,
   ) {
     _populateElementRecorderMap([
       ContainerRecorder(keyGenerator),
       TextElementRecorder(keyGenerator),
       EditableTextRecorder(keyGenerator),
       InputDecoratorRecorder(keyGenerator),
-      ImageRecorder(keyGenerator),
+      ImageRecorder(
+        keyGenerator,
+        imageDownscaling: imageDownscaling,
+        internalLogger: internalLogger,
+      ),
       CustomPaintRecorder(keyGenerator),
       PrivacyRecorder(keyGenerator),
     ]);